Assignment 1.2
Problem Statement
1. How many ways are there to call a function in R?


Answer : A function is a piece of code written to carry out a specified task. Different methods to call a function in R are:- 
Call( name, …..) # returns an unevaluated function call 
do.call # for calling a function by name and argument list 
Recall # for recursive calling of functions

2.	Is the below statement true?

•	The lazy evaluation of a function means, the argument is evaluated only if it is evaluated only if it is used inside the body of the function.

Answer : True


3.	Mention true or false for below statements:

a.	Insights driven from descriptive analytics is not meaningful. : 
Answer	: False
b.	The number of values in each Elements of a list, should be equal. : 
Answer	: False
c.	The datasets are not stored in memory of the computer using R. : 
Answer	: True
d.	Data frames and matrices are two dimensional however the array is multidimensional. 
Answer	: True